Revert "fix(preview-service): only create a puppeteer client per app, not per…"
This reverts commit d5c9e5e207aad4869f7e2b9562f3da75173ba7c0.
