+++ packages/preview-service/tests/helpers/helpers.ts
@@ -3,21 +3,13 @@ import type { Knex } from 'knex'
 import http from 'http'
 import type { AddressInfo } from 'net'
 import { getPostgresConnectionString } from '@/utils/env.js'
-import { PuppeteerClient } from '@/clients/puppeteer.js'
 
-export const startAndWaitOnServers = async (deps: {
-  db: Knex
-  puppeteerClient: PuppeteerClient
-}) => {
+export const startAndWaitOnServers = async (deps: { db: Knex }) => {
   let serverAddress: string | AddressInfo | null = null
   let metricsServerAddress: string | AddressInfo | null = null
 
-  const { db, puppeteerClient } = deps
-  const { app, server, metricsServer } = startServer({
-    db,
-    puppeteerClient,
-    serveOnRandomPort: true
-  })
+  const { db } = deps
+  const { app, server, metricsServer } = startServer({ db, serveOnRandomPort: true })
   server.on('listening', () => {
     serverAddress = server.address()
   })
