+++ packages/preview-service/tests/helpers/testExtensions.ts
@@ -4,7 +4,6 @@ import { getTestDb } from '#/helpers/testKnexClient.js'
 import { startAndWaitOnServers } from '#/helpers/helpers.js'
 import type { Knex } from 'knex'
 import { Server } from 'http'
-import { PuppeteerClient } from '@/clients/puppeteer.js'
 
 export interface AcceptanceTestContext {
   context: {
@@ -79,20 +78,12 @@ export const e2eTest = test.extend<E2ETestContext>({
       const dbName = inject('dbName')
       // equivalent of beforeEach
       const db = await getTestDb(dbName).transaction()
-      const puppeteerClient: PuppeteerClient = {
-        //FIXME this is a stub for the moment
-        loadPageAndEvaluateScript: async () => {},
-        dispose: async () => {}
-      }
-      const { server, metricsServer } = await startAndWaitOnServers({
-        db,
-        puppeteerClient
-      })
+      const { server, metricsServer } = await startAndWaitOnServers({ db })
 
       // schedule the cleanup. Runs regardless of test status, and runs after afterEach.
       onTestFinished(async () => {
-        if (server) await stopServer({ server, puppeteerClient })
-        if (metricsServer) await stopServer({ server: metricsServer, puppeteerClient })
+        if (server) stopServer({ server })
+        if (metricsServer) stopServer({ server: metricsServer })
         if (db) await db.rollback()
       })
